move GeorgeIaLogL to ia.py
