+++ jesi/likelihoods/des5y.py
@@ -1,8 +1,7 @@
 from numpy import loadtxt, argsort, sqrt, fill_diagonal
-from jax.numpy import array
 import pandas as pd
 from pathlib import Path
-from jesi.likelihoods.ia import IaLogL
+from jesi.likelihoods.ia import IaLogL, GeorgeIaLogL
 
 
 # data loading stolen from Toby
@@ -21,20 +20,4 @@ logl = IaLogL(df, cov, 'MU')
 # george fiddle
 des_id = 10.0
 george_mask = df['IDSURVEY'] != des_id
-
-
-class GeorgeIaLogL(IaLogL):
-    requirements = {'h0_dl_over_c', 'delta_mb'}
-
-    def __init__(self, george_mask, *args, **kwargs):
-        self.george_mask = array(george_mask.to_numpy())
-        super().__init__(*args, **kwargs)
-
-    def _y(self, params, cosmology):
-        # offset is the george offset
-        y = super()._y(params, cosmology)
-        y -= params['delta_mb'] * self.george_mask
-        return y
-
-
 logloffset = GeorgeIaLogL(george_mask, df, cov, mb_column='MU', z_cutoff=0.0)
